refactor: extract settings runway threshold slider
